Move summary to cpp (cherry picked from commit 6db2c9a)
